Move make_run_report to swebench/harness/reporting for run_evaluation and run_evaluation_modal
