+++ swebench/harness/modal_eval/run_evaluation_modal.py
@@ -13,8 +13,8 @@ import traceback
 
 from dataclasses import dataclass
 from pathlib import Path
-from swebench.harness.constants import KEY_INSTANCE_ID
 from swebench.harness.docker_build import setup_logger
+from swebench.harness.reporting import make_run_report
 from swebench.harness.utils import EvaluationError
 from typing import cast
 
@@ -195,109 +195,12 @@ class ModalSandboxRuntime:
             )
             .workdir("/testbed/")
         )
-    
-def make_run_report(
-        predictions: dict,
-        full_dataset: list,
-        run_id: str
-    ) -> Path:
-    """
-    Make a final evaluation and run report of the instances that have been run.
 
-    Args:
-        predictions (dict): Predictions dict generated by the model
-        full_dataset (list): List of all instances
-        run_id (str): Run ID
-    
-    Returns:
-        Path to report file
-    """
-    # Sets to store IDs of different outcomes
-    completed_ids = set()
-    resolved_ids = set()
-    error_ids = set()
-    unresolved_ids = set()
-    incomplete_ids = set()
-    empty_patch_ids = set()
-
-    for instance in full_dataset:
-        instance_id = instance[KEY_INSTANCE_ID]
-
-        # Instances that were not submitted
-        if instance_id not in predictions:
-            incomplete_ids.add(instance_id)
-            continue
-        
-        # Instances with empty patches
-        prediction = predictions[instance_id]
-        if prediction.get("model_patch", None) in ["", None]:
-            empty_patch_ids.add(instance_id)
-            continue
-
-        # Instances that errored
-        log_dir = get_log_dir(predictions[instance_id], run_id, instance_id)
-        report_file = log_dir / "report.json"
-        if not report_file.exists():
-            error_ids.add(instance_id)
-            continue
-
-        # Instance completed successfully
-        completed_ids.add(instance_id)
-        try:
-            report = json.loads(report_file.read_text())
-            if report[instance_id]["resolved"]:
-                resolved_ids.add(instance_id)
-            else:
-                unresolved_ids.add(instance_id)
-        except Exception as e:
-            print(f"{instance_id}: error loading report.json: {e}")
-            error_ids.add(instance_id)
-
-    # Print final report
-    dataset_ids = {i[KEY_INSTANCE_ID] for i in full_dataset}
-    print(f"Total instances: {len(full_dataset)}")
-    print(f"Instances submitted: {len(set(predictions.keys()) & dataset_ids)}")
-    print(f"Instances completed: {len(completed_ids)}")
-    print(f"Instances incomplete: {len(incomplete_ids)}")
-    print(f"Instances resolved: {len(resolved_ids)}")
-    print(f"Instances unresolved: {len(unresolved_ids)}")
-    print(f"Instances with empty patches: {len(empty_patch_ids)}")
-    print(f"Instances with errors: {len(error_ids)}")
-
-    # Write report to file
-    report = {
-        "total_instances": len(full_dataset),
-        "submitted_instances": len(predictions),
-        "completed_instances": len(completed_ids),
-        "resolved_instances": len(resolved_ids),
-        "unresolved_instances": len(unresolved_ids),
-        "empty_patch_instances": len(empty_patch_ids),
-        "error_instances": len(error_ids),
-        "completed_ids": list(sorted(completed_ids)),
-        "incomplete_ids": list(sorted(incomplete_ids)),
-        "empty_patch_ids": list(sorted(empty_patch_ids)),
-        "submitted_ids": list(sorted(predictions.keys())),
-        "resolved_ids": list(sorted(resolved_ids)),
-        "unresolved_ids": list(sorted(unresolved_ids)),
-        "error_ids": list(sorted(error_ids)),
-        "schema_version": 2,
-    }
-
-    report_file = Path(
-        list(predictions.values())[0]["model_name_or_path"].replace("/", "__")
-        + f".{run_id}"
-        + ".json"
-    )
-
-    with open(report_file, "w") as f:
-        print(json.dumps(report, indent=4), file=f)
-
-    print(f"Report written to {report_file}")
-    return report_file
 
 def get_log_dir(pred: dict, run_id: str, instance_id: str) -> Path:
     model_name_or_path = cast(str, pred.get("model_name_or_path", "None").replace("/", "__"))
     return RUN_EVALUATION_LOG_DIR / run_id / model_name_or_path / instance_id
+
 
 @app.function(
     image=swebench_image.add_local_file(
@@ -415,7 +318,7 @@ def run_instance_modal(
         report = get_eval_report(
             test_spec=test_spec,
             prediction=pred,
-            log_path=test_output_path,
+            test_log_path=test_output_path,
             include_tests_status=True,
         )
         logger.info(
@@ -464,6 +367,7 @@ def run_instance_modal(
             log_dir=log_dir,
             errored=True,
         )
+
 
 def run_instances_modal(
         predictions: dict,
