+++ tests/test_evaluation.py
@@ -27,8 +27,8 @@ def test_make_run_report(tmpdir) -> None:
                 }
             },
             [TEST_INSTANCE],
+            "test",
             client,
-            "test"
         )
         assert output_path.is_file()
         report = json.loads(output_path.read_text())
